fix: organize profile
