+++ resources/js/pages/settings/profile/useHooks.ts
@@ -1,0 +1,46 @@
+import { BreadcrumbItem, SharedData } from '@/types';
+import { useForm, usePage } from '@inertiajs/react';
+import { FormEventHandler } from 'react';
+
+type ProfileForm = {
+    first_name: string;
+    middle_name: string;
+    last_name: string;
+    email: string;
+};
+
+const breadcrumbs: BreadcrumbItem[] = [
+    {
+        title: 'Profile settings',
+        href: '/settings/profile',
+    },
+];
+
+export const useHooks = () => {
+    const { auth } = usePage<SharedData>().props;
+
+    const { data, setData, patch, errors, processing, recentlySuccessful } = useForm<Required<ProfileForm>>({
+        first_name: auth.user.first_name,
+        middle_name: auth.user.middle_name,
+        last_name: auth.user.last_name,
+        email: auth.user.email,
+    });
+
+    const submit: FormEventHandler = (e) => {
+        e.preventDefault();
+
+        patch(route('profile.update'), {
+            preserveScroll: true,
+        });
+    };
+    return {
+        breadcrumbs,
+        auth,
+        data,
+        setData,
+        errors,
+        processing,
+        recentlySuccessful,
+        submit,
+    };
+};
